backend/classification-service/pyproject.toml
[project]
name = "classification-service"
version = "0.1.0"
description = ""
authors = [
    {name = "1337kid",email = "45532566+1337kid@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "pandas (>=3.0.1,<4.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "joblib (>=1.5.3,<2.0.0)",
    "google-cloud-storage (>=3.9.0,<4.0.0)",
    "google-auth (>=2.49.0,<3.0.0)",
    "fastapi (>=0.110,<1.0.0)",
    "uvicorn[standard] (>=0.29,<1.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "supabase (>=2.28.3,<3.0.0)"
]

[tool.poetry]
package-mode = false